pyproject.toml
[project]
name = "polygonio-ingestion"
version = "0.1.0"
description = "Polygon.io data ingestion → Parquet lake (minute/day + corp actions)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pandas>=2.2","pyarrow>=16","numpy","tqdm","tenacity>=8","loguru>=0.7","typer>=0.12","polygon-api-client>=1.13","python-dotenv>=1.0"]

[project.scripts]
poly = "polygon_ingest.cli:app"

[tool.pytest.ini_options]
pythonpath = ["src"]
